backend/src/main/resources/etl/requirements.txt
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudscraper==1.2.71
h11==0.16.0
html5lib==1.1
idna==3.11
lxml==4.9.4
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soccerdata==1.8.7
sortedcontainers==2.4.0
tenacity==9.1.2
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
undetected-chromedriver==3.5.5
unicode==2.9
Unidecode==1.4.0
urllib3==2.6.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
